consolidate collection manager service into rest api service
