--- app/src/app/services/connectors/collection-manager/collection-manager-connector.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { CollectionManagerConnectorService } from './collection-manager-connector.service';
-
-describe('CollectionManagerConnectorService', () => {
-  let service: CollectionManagerConnectorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(CollectionManagerConnectorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
